+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Improvements to the user experience
 
 - Remove the redundant login screen - fixed redirect ([#1360, #1463](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1360, https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1463))
+- Fix: [Lobby] no feedback for wrong password ([1440](https://git.opentalk.dev/opentalk/frontend/web/web-app/-/issues/1440))
 ### Bug Fixes
 
 - Fix push-to-talk functionality([#1344]